requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.6.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
blobfile==2.0.1
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.2
certifi==2021.10.8
charset-normalizer==2.1.1
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.0.3
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.6.1
debugpy==1.6.5
decorator==5.1.1
Deprecated==1.2.14
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.104.1
ffmpeg==1.4
ffprobe-python==1.0.3
filelock==3.9.0
Flask==2.0.2
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.9.2
google-auth==2.23.4
googleapis-common-protos==1.61.0
grpcio==1.59.2
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.14.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.20.1
ipython==8.10.0
itsdangerous==2.0.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==7.4.8
jupyter_core==5.1.3
kiwisolver==1.4.4
kubernetes==28.1.0
langchain==0.0.327
langsmith==0.0.56
llvmlite==0.41.0
lxml==4.9.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numba==0.58.0
numpy==1.24.1
oauthlib==3.2.2
onnxruntime==1.16.1
openai==0.28.1
openai-whisper @ git+https://github.com/openai/whisper.git@0a60fcaa9b86748389a656aa013c416030287d47
openpyxl==3.0.9
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
overrides==7.4.0
packaging==23.0
pandas==1.3.4
pandas-stubs==1.2.0.35
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.12
platformdirs==2.6.2
plotly==5.12.0
posthog==3.0.2
prompt-toolkit==3.0.36
protobuf==4.25.0
psutil==5.9.4
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pycryptodomex==3.17
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pyee==11.0.1
Pygments==2.15.0
pyparsing==3.0.9
pypdf==3.17.0
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==0.19.2
python-ffmpeg==2.0.4
pytz==2022.7.1
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-learn==1.2.0
scipy==1.10.0
semantic-version==2.10.0
setuptools-rust==1.7.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.22
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.8.1
transformers==4.30.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==1.26.13
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.5
websocket-client==1.6.4
websockets==12.0
Werkzeug==2.0.2
wrapt==1.15.0
yarg==0.1.9
yarl==1.8.2
youtube-transcript-api==0.6.1
yt-dlp==2023.10.13
zipp==3.17.0
